+++ packages/terra-application-links/tests/jest/ApplicationTabs.test.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
+import { BrowserRouter } from 'react-router-dom';
 import ApplicationTabs from '../../src/tabs/ApplicationTabs';
 import testLinkConfig from './testLinkConfig';
 import testLinksWithIconsConfig from './testLinksWithIconsConfig';
@@ -9,6 +10,19 @@ import testLinksWithIconsConfig from './testLinksWithIconsConfig';
 it('should render ApplicationTabs with links and alignment', () => {
   const wrapper = enzyme.shallow(<div><ApplicationTabs links={testLinkConfig} alignment="start" /></div>);
   expect(wrapper).toMatchSnapshot();
+});
+
+it('should trigger onTabClick handler', () => {
+  const onTabClickHandler = jest.fn();
+  const wrapper = enzymeIntl.mountWithIntl(
+    <BrowserRouter>
+      <ApplicationTabs links={testLinkConfig} alignment="start" onTabClick={onTabClickHandler} />
+    </BrowserRouter>,
+  );
+
+  wrapper.find('.tab').first().simulate('click');
+
+  expect(onTabClickHandler).toHaveBeenCalled();
 });
 
 it('should render ApplicationTabs with icons', () => {