+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 6.99.0 - (April 23, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 6.98.0 - (April 18, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 6.97.1 - (April 17, 2024)
+
+* Fixed
+  * Set aria-haspopup value to boolean.
+
 ## 6.97.0 - (April 4, 2024)
 
 * Changed