+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.114.1 - (April 23, 2024)
+
+* Fixed
+  * Fixed the accessibility issue to allow announcement as a single group label.
+
 ## 4.114.0 - (April 4, 2024)
 
 * Changed