+++ packages/terra-file-path/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 1.17.0 - (April 23, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.16.0 - (April 18, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.15.0 - (April 17, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.14.0 - (April 4, 2024)
 
 * Changed