+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 6.94.0 - (April 23, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 6.93.0 - (April 18, 2024)
+
+* Added
+  * Added test for `terra-table` to cover exception occurring with collapsible sections and single row select mode.
+
+## 6.92.0 - (April 17, 2024)
+
+* Changed
+  * Removed dialog role to fix axe a11y violation.
+
 ## 6.91.0 - (April 4, 2024)
 
 * Changed