+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.46.1 - (April 17, 2024)
+
+* Fixed
+ * Fixed Swedish translations for headers with new translations used elsewhere.
+
 ## 4.46.0 - (April 4, 2024)
 
 * Changed